pyproject.toml
[tool.poetry]
name = "SportTracker"
version = "1.37.0"
description = "Self-hosted sport data tracking server."
license = "GPL-3.0-or-later"
authors = ["deadlocker8"]

[[tool.poetry.source]]
name = "TheCodeLabs"
url = "https://pypi.thecodelabs.de"
priority = "explicit"

[tool.poetry.dependencies]
python = "^3.12"
thecodelabs-baseutils = { version = "*", source = "TheCodeLabs" }
thecodelabs-flaskutils = { version = "*", source = "TheCodeLabs" }
flask = "3.1.0"
gevent = "24.11.1"
click = "8.1.8"
Flask-SQLAlchemy = "3.1.1"
Flask-Pydantic = "0.12.0"
Flask-Login = "0.6.3"
Flask-Bcrypt = "1.0.1"
Flask-Migrate = "4.1.0"
flask-babel = "4.0.0"
python-dateutil = "2.9.0.post0"
psycopg2 = "2.9.10"
gpxpy = "1.6.2"
faker = "37.1.0"
pillow = "11.2.1"
fitdecode = "0.10.0"

[tool.poetry.group.dev.dependencies]
ruff = "0.11.2"
pre-commit = "4.2.0"
mypy = "1.15.0"
pytest = "8.3.5"
pytest-selenium = "4.1.0"
selenium = "4.30.0"

types-python-dateutil = "*"
types-requests = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]

[tool.ruff.format]
quote-style = "single"

[tool.mypy]
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "flask_babel.*",
    "flask_bcrypt.*",
    "flask_login.*",
    "flask_pydantic.*",
    "flask_migrate.*",
    "TheCodeLabs_BaseUtils.*",
    "TheCodeLabs_FlaskUtils.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = "sporttracker"
